requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
